moved secrets for AAT to jenkins file
