+++ Jenkinsfile_nightly
@@ -22,8 +22,6 @@ def secrets = [
         secret('idam-redirect-uri', 'IA_IDAM_REDIRECT_URI'),
         secret('idam-secret', 'IA_IDAM_SECRET'),
         secret('wa-idam-client-secret', 'WA_IDAM_CLIENT_SECRET'),
-        secret('wa-test-law-firm-a-username', 'TEST_WA_LAW_FIRM_USERNAME'),
-        secret('wa-test-law-firm-a-password', 'TEST_WA_LAW_FIRM_PASSWORD'),
     ],
 
   's2s-${env}' : [
