+++ src/testUtils/java/uk/gov/hmcts/reform/wataskmanagementapi/services/IdamServiceApi.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskmanagementapi.clients;
+package uk.gov.hmcts.reform.wataskmanagementapi.services;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
